agent-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
anthropic
mem0ai
dateparser
requests
python-dotenv